contact information/requirements.txt
tkinter
pillow
sqlite3